// src/enum/config.enum.ts
export enum ConfigEnum {
  // 数据库
  DB_TYPE = 'DB_TYPE',
  DB_HOST = 'DB_HOST',
  DB_PORT = 'DB_PORT',
  DB_DATABASE = 'DB_DATABASE',
  DB_USERNAME = 'DB_USERNAME',
  DB_PASSWORD = 'DB_PASSWORD',
  DB_SYNC = 'DB_SYNC',
  DB_LOGGING = 'DB_LOGGING',
  // NestJS 程序
  APP_PORT = 'APP_PORT',
  // winston日志
  LOG_FILE = 'LOG_FILE',
  LOG_DIRNAME = 'LOG_DIRNAME',
  // JWT 密钥
  JWT_SECRET = 'JWT_SECRET',
}